% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetFakeRegData.R
\name{getFakeRegData}
\alias{getFakeRegData}
\title{Provide dataframe of fake registry data}
\usage{
getFakeRegData()
}
\value{
regData data frame
}
\description{
Provides a dataframe containing built-in data (and not a registry) for demo
purposes
}
