% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetRegData.R
\name{getRegData}
\alias{getRegData}
\title{Provide dataframe of registry data}
\usage{
getRegData(registryName)
}
\arguments{
\item{registryName}{String providing the current registryName}
}
\value{
regData data frame
}
\description{
Provides a dataframe containing data from a registry
}
